calculations now read properly .dat files and are crrectly working

+++ lib/node.h
@@ -1,5 +1,5 @@
 
 typedef struct {
     char *filename;
-    long result;
+    long long result;
 } _node;
